Tested com.health.gui, applied some minor tweaks
